+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - Reduced the number of emails per mailbox in the test data !3330 @alex.simm
 - Remove unused code !3340 @alex.simm
 - Fixed errors in the seed script !3355 @alex.simm
+- Uploaded files without a usage type are deleted in the nightly maintenance #1182 !3363 @alex.simm
 
 ## Development documentation